Rename the building_id field to nycdb_building_id
I was using the building_id field to identify visits. This was a bad design
choice that I didn't catch, because it doesn't allow for notekeeping on properties
that aren't in the nycdb. I'll be changing that. So i renamed the respective fields
to make the purpose of the field clear.